Fix PicklingError under certain platforms
test_fibonacci_multiprocessing was failing under macos (darwin) with the
following error:

  _pickle.PicklingError: Can't pickle <function rebuild_connection at ...>: import of module 'multiprocessing.connection' failed

This is due to the fact that functions are only picklable if they are
defined at the top-level of a module, therefore a @staticmethod of
DaemonTask would not work under certain platforms such as darwin, where
tasks are `spawn`ed (not `fork`ed) for which pickling are necessary.

See e.g. https://stackoverflow.com/a/8805244